Parsing kQueries without leaking memory

+++ synapse/src/execution_plan/meta.cpp
@@ -39,8 +39,8 @@ void EPMeta::update(const EPLeaf &leaf, const EPNode *new_node,
   nodes++;
 }
 
-float EPMeta::get_bdd_progress() const {
-  return processed_nodes.size() / static_cast<float>(total_bdd_nodes);
+double EPMeta::get_bdd_progress() const {
+  return processed_nodes.size() / static_cast<double>(total_bdd_nodes);
 }
 
 void EPMeta::update_total_bdd_nodes(const BDD *bdd) { total_bdd_nodes = bdd->size(); }
